Album information functions on a seperate file
