+++ src/releases.js
@@ -1,87 +1,6 @@
-const axios = require('axios');
 const spotify = require('../api/spotify-properties');
 const { checkTodayRelease, sleep } = require('../utils/utils');
-const Album = require('./components/Album');
-
-async function getFullAlbumDetails(href) {
-    let result;
-    try {
-        result = await axios.get(href, spotify.getAuthOptions());
-    } catch (err) {
-        await sleep(err.response.headers['retry-after'] * 1000);
-        return;
-    }
-
-    const fullAlbumDetails = result.data;
-    return fullAlbumDetails;
-}
-
-function getAuthors(album) {
-    const albumArtists = album.artists.filter(artist => {
-        if (!album.name.includes(artist.name)) {
-            return artist.name
-        }
-    }).map(artist => artist.name);
-    var authors = '';
-    if (albumArtists.length !== 0) {
-        for (var i = 0; i < albumArtists.length - 1; i++) {
-            authors += albumArtists[i] + ' & ';
-        }
-        authors += albumArtists[albumArtists.length - 1];
-    } else {
-        for (var i = 0; i < album.artists.length - 1; i++) {
-            authors += album.artists[i].name + ' & ';
-        }
-        authors += album.artists[album.artists.length - 1].name;
-    }
-
-    return authors;
-}
-
-function getFeaturedArtists(fullAlbumDetails, artists) {
-    const trackArtists = fullAlbumDetails.tracks.items[0].artists.filter(artist => {
-        if (!artists.includes(artist.name) && !fullAlbumDetails.name.includes(artist.name)) {
-            return artist.name
-        }
-    }).map(artist => artist.name);
-
-    let featuredArtists = '';
-    if (trackArtists.length > 0) {
-        featuredArtists = '(ft. ';
-        for (var i = 0; i < trackArtists.length - 1; i++) {
-            featuredArtists += trackArtists[i] + ' & ';
-        }
-        featuredArtists += trackArtists[trackArtists.length - 1] + ')';
-    }
-
-    return featuredArtists;
-}
-
-function getTracklist(fullAlbumDetails, titleArtists, title) {
-    let tracklist = '';
-    const tracks = fullAlbumDetails.tracks.items;
-
-    for (var i = 0; i < tracks.length; i++) {
-        tracklist += (i + 1) + '. ' + tracks[i].name;
-        if (tracks[i].artists.length > 1 && !tracks[i].name.toLowerCase().includes('remix')) {
-            const artistNames = tracks[i].artists.filter(artist => {
-                if (!titleArtists.includes(artist.name) && !title.includes(artist.name)) {
-                    return artist.name
-                }
-            }).map(artist => artist.name);
-            if (artistNames.length !== 0) {
-                tracklist += ' (w/ ';
-                for (var j = 1; j < artistNames.length - 1; j++) {
-                    tracklist += artistNames[j] + ' & ';
-                }
-                tracklist += artistNames[artistNames.length - 1] + ')';
-            }
-        }
-        tracklist += '\n';
-    }
-
-    return tracklist;
-}
+const getAlbumInfos = require('./releases/infoReleases');
 
 async function getNewestReleases(artistId) {
     let newestReleases = [];
@@ -112,60 +31,6 @@ async function getLatestReleases(artistId) {
     } else {
         return [];
     }
-}
-
-async function getAlbumInfos(albums) {
-    let latestReleases = [];
-    for (var i = 0; i < albums.length; i++) {
-        let fullAlbumDetails = await getFullAlbumDetails(albums[i].href);
-
-        while (fullAlbumDetails === undefined) {
-            fullAlbumDetails = await getFullAlbumDetails(albums[i].href);
-        }
-
-        const artists = getAuthors(albums[i]);
-        const nameAlbum = albums[i].name;
-
-        if (!latestReleases.some(release => release.nameAlbum.toLowerCase().includes(nameAlbum.toLowerCase()))) {
-            let link;
-            let tracklist;
-            let featuredArtists;
-            let album;
-
-            if (fullAlbumDetails.total_tracks > 1) {
-                const title = artists + '\n' + nameAlbum;
-                tracklist = getTracklist(fullAlbumDetails, artists, title);
-                link = albums[i].external_urls.spotify;
-                album = new Album({
-                    nameAlbum: nameAlbum,
-                    artists: artists,
-                    featArtists: '',
-                    label: fullAlbumDetails.label,
-                    releaseDate: albums[i].release_date,
-                    tracklist: tracklist,
-                    coverArt: fullAlbumDetails.images[0].url,
-                    spotifyLink: link
-                })
-            } else {
-                featuredArtists = getFeaturedArtists(fullAlbumDetails, artists);
-                link = fullAlbumDetails.tracks.items[0].external_urls.spotify;
-                album = new Album({
-                    nameAlbum: nameAlbum,
-                    artists: artists,
-                    featArtists: featuredArtists,
-                    label: fullAlbumDetails.label,
-                    releaseDate: albums[i].release_date,
-                    tracklist: '',
-                    coverArt: fullAlbumDetails.images[0].url,
-                    spotifyLink: link
-                })
-            }
-
-            latestReleases.push(album);
-        }
-    }
-
-    return latestReleases;
 }
 
 async function getLatestAlbumObjects(artistId) {
